Move examples into folder
